[FIX] queue_job: recompute with user
When queue_job is not used, recomputing is done with the same user that
did the changes.
queue_job changes that and the superuser is used instead. That causes
problems. For example, when mail are sent, if the superuser has no mail
address.

After this change, the original user is used instead of the superuser.